Replacing fonts extensions
